correction: added verify operator method

+++ basicCalculator/src/main/java/org/example/calculator/record/ResultRecord.java
@@ -1,4 +1,4 @@
 package org.example.calculator.record;
 
-public record ResultRecord(int result) {
+public record ResultRecord(double result) {
 }
